fix(core): avoid perf.now for wake-up detection
performance.now might ignore computer sleep time, which is a
false-negative.

In comparison, Date.now() is affected by system clock change, which
causes a false-positive.

In this context, a false-positive is more acceptable compared to a
false-negative, so use of Date.now is preferred.

Side note: Temporal.Now polyfill uses Date.now internally, so that
can be used as well.